btrfs: use nocow_end for the loop iteration in run_delalloc_cow
When run_delalloc_cow allocates an ordered extent for an actual NOCOW
range, it uses the nocow_end variable calculated based on the current
offset and the nocow_args.num_bytes value returned from
can_nocow_file_extent for all the actual I/O, but the loop iteration
then resets cur_offset to extent_end, which caused me a lot of
confusion.

It turns out that nocow_end is based of the minimum of the extent end
and the range end, and thus actually works perfectly fine for the loop
iteration, but using a different variable here from the actual I/O
submission is confusing.  Switch to using nocow_end adjusted by the off
by one to make it an exclusive range.

Reviewed-by: Boris Burkov <boris@bur.io>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: David Sterba <dsterba@suse.com>