soc: qcom: Add driver to read secondary bootloader (XBL) log
Qualcomm secondary bootloader (XBL) boot log holds information to
identify various firmware configuration currently set on the SoC.
The XBL log is stored in a predefined reserved memory region.

This drivers provides a way to print XBL logs on the console. To
do so, it provides a debugfs entry which captures the logs stored
in this reserved memory region. This entry can now be used to read
and print the XBL logs to console.

User can use the below command to print XBL log to console:
        cat /sys/kernel/debug/xbl_log

Signed-off-by: Ninad Naik <quic_ninanaik@quicinc.com>