dm: set the correct discard_sectors limit
Since commit 0034af036554 ("block: make
/sys/block/<dev>/queue/discard_max_bytes writeable") there are two
max_discard_sectors limits, one provided by the driver and one set by the
user to optionally reduce the size below the hardware or driver limits.
Usage of the extra hw limits has been a bit convoluted and both were
set by the same driver API, leading to potential overrides of the user
setting by the driver updating the limits.

With the new atomic queue limits API the driver should only set the hw
limit, but I forgot to convert dm over to that as it was already using
a scheme where the queue_limits are passed around.  Fix dm to update
the max_hw_discard_sectors limits.

Note that this still leaves the non-hw update in place, which should
be removed to not override the user settings.  As that is a behavior
change I do not want to do it at the very end of the merge window.

This fixes a regression where dm bio prison v1 warns about exceeding
the discard bio size when running xfstests generic/500.

Fixes: 8e0ef4128694 ("dm: use queue_limits_set")
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Mike Snitzer <snitzer@kernel.org>