block: implement async write zero pages command
Add a command that writes the zero page to the drive. Apart from passing
the zero page instead of actual data it uses the normal write path and
doesn't do any further acceleration, nor it requires any special
hardware support. The indended use is to have a fallback when
BLOCK_URING_CMD_WRITE_ZEROES is not supported.

Signed-off-by: Pavel Begunkov <asml.silence@gmail.com>
Link: https://lore.kernel.org/r/c465430b0802ced71d22f548587f2e06951b3cd5.1725621577.git.asml.silence@gmail.com
Signed-off-by: Jens Axboe <axboe@kernel.dk>