arm64: dts: rockchip: add msi-parent for pcie3x4_ep on rk3588
Add msi-parent for the pcie3x4_ep PCI endpoint node.

The pcie3x4_ep node should use the same msi-parent as the pcie3x4 node
(which represents the PCIe controller running in Root Complex mode).

The GIC ITS can be used to trigger an IRQ on the endpoint when any of
the endpoint's PCI BARs are written to by the host[1].

[1] https://lore.kernel.org/linux-pci/20241116-ep-msi-v8-0-6f1f68ffd1bb@nxp.com/

Signed-off-by: Niklas Cassel <cassel@kernel.org>
Link: https://lore.kernel.org/r/20241120171048.2839621-2-cassel@kernel.org
Signed-off-by: Heiko Stuebner <heiko@sntech.de>